+++ packages/cubejs-api-gateway/src/gateway.ts
@@ -119,7 +119,7 @@ class ApiGateway {
   protected readonly contextToApiScopesFn: ContextToApiScopesFn;
 
   protected readonly contextToApiScopesDefFn: ContextToApiScopesFn =
-    async () => ['liveliness', 'graphql', 'meta', 'data'];
+    async () => ['graphql', 'meta', 'data'];
 
   protected readonly checkAuthMiddleware: CheckAuthMiddlewareFn;
 
@@ -181,11 +181,10 @@ class ApiGateway {
     ];
 
     /** **************************************************************
-     * Liveliness scope                                              *
+     * No scope                                              *
      *************************************************************** */
 
     // @todo Should we pass requestLoggerMiddleware?
-    // @todo Should we add scope assert here?
 
     const guestMiddlewares = [];
 
@@ -193,7 +192,7 @@ class ApiGateway {
     app.get('/livez', guestMiddlewares, cachedHandler(this.liveness));
 
     /** **************************************************************
-     * Graphql scope                                                 *
+     * graphql scope                                                 *
      *************************************************************** */
 
     app.use(
@@ -239,7 +238,7 @@ class ApiGateway {
     );
 
     /** **************************************************************
-     * Data scope                                                    *
+     * data scope                                                    *
      *************************************************************** */
 
     app.get(`${this.basePath}/v1/load`, userMiddlewares, (async (req, res) => {
@@ -302,6 +301,10 @@ class ApiGateway {
       });
     }));
 
+    /** **************************************************************
+     * meta scope                                                    *
+     *************************************************************** */
+
     app.get(
       `${this.basePath}/v1/meta`,
       userMiddlewares,
@@ -320,6 +323,7 @@ class ApiGateway {
       }
     );
 
+    // Used by Rollup Designer
     app.post(
       `${this.basePath}/v1/pre-aggregations/can-use`,
       [
@@ -352,7 +356,7 @@ class ApiGateway {
     );
 
     /** **************************************************************
-     * Jobs scope                                                    *
+     * jobs scope                                                    *
      *************************************************************** */
 
     app.get(
@@ -1974,7 +1978,7 @@ class ApiGateway {
           );
         } else {
           scopes.forEach((p) => {
-            if (['liveliness', 'graphql', 'meta', 'data', 'jobs'].indexOf(p) === -1) {
+            if (['graphql', 'meta', 'data', 'jobs'].indexOf(p) === -1) {
               throw new Error(
                 `A user-defined contextToApiScopes function returns a wrong scope: ${p}`
               );
@@ -2007,7 +2011,7 @@ class ApiGateway {
       throw new CubejsHandlerError(
         403,
         'Forbidden',
-        `Api scope is missed: ${scope}`
+        `API scope is missing: ${scope}`
       );
     }
   }