+++ packages/cubejs-backend-shared/src/platform.ts
@@ -1,6 +1,7 @@
 import { spawnSync } from 'child_process';
+import process from 'process';
 import { internalExceptions } from './errors';
-import { displayCLIWarning } from './cli';
+import { displayCLIWarning, displayCLIWarningOnce } from './cli';
 
 export function detectLibc() {
   if (process.platform !== 'linux') {
@@ -48,4 +49,59 @@ export function detectLibc() {
   displayCLIWarning('Unable to detect what host library is used as libc, continue with gnu');
 
   return 'gnu';
+}
+
+type IsNativeSupportedResult = true | {
+  reason: string
+};
+
+export function isNativeSupported(): IsNativeSupportedResult {
+  if (process.platform === 'linux') {
+    if (detectLibc() === 'musl') {
+      displayCLIWarningOnce(
+        'is-native-supported',
+        'Unable to load native extension. You are using a Linux distro with Musl which is not supported. Read more: ' +
+        'https://github.com/cube-js/cube/blob/master/packages/cubejs-backend-native/README.md#supported-architectures-and-platforms'
+      );
+
+      return {
+        reason: 'You are using linux distro with Musl which is not supported'
+      };
+    }
+  }
+
+  return true;
+}
+
+export enum LibraryExistsResult {
+  // We are sure that required library doesnt exist on system
+  None,
+  // We are sure that required library exists
+  Exists,
+  UnableToCheck
+}
+
+export function libraryExists(libraryName: string): LibraryExistsResult {
+  if (process.platform === 'linux') {
+    try {
+      const { status, output } = spawnSync('ldconfig', ['-v'], {
+        encoding: 'utf8',
+        // Using pipe to protect unexpect STDERR output
+        stdio: 'pipe'
+      });
+      if (status === 0) {
+        if (output.join(' ').includes(libraryName)) {
+          return LibraryExistsResult.Exists;
+        }
+
+        return LibraryExistsResult.None;
+      } else {
+        return LibraryExistsResult.UnableToCheck;
+      }
+    } catch (e: any) {
+      internalExceptions(e);
+    }
+  }
+
+  return LibraryExistsResult.UnableToCheck;
 }