+++ packages/cubejs-ksql-driver/src/KsqlDriver.ts
@@ -72,7 +72,7 @@ export class KsqlDriver extends BaseDriver implements DriverInterface {
    * Returns default concurrency value.
    */
   public static getDefaultConcurrency(): number {
-    return 2;
+    return 1;
   }
 
   protected readonly config: KsqlDriverOptions;