+++ packages/cubejs-schema-compiler/src/scaffolding/formatters/BaseSchemaFormatter.ts
@@ -99,21 +99,21 @@ export abstract class BaseSchemaFormatter {
   }
 
   protected eligibleIdentifier(name: string) {
-    return !!name.match(/^[a-z0-9_]+$/);
+    return !!name.match(/^[a-z0-9_]+$/i);
   }
 
   public schemaDescriptorForTable(tableSchema: TableSchema, schemaContext: SchemaContext = {}) {
     const table = `${
       tableSchema.schema?.length ? `${this.escapeName(tableSchema.schema)}.` : ''
     }${this.escapeName(tableSchema.table)}`;
-    
+
     const { dataSource, ...contextProps } = schemaContext;
-      
+
     let dataSourceProp = {};
     if (dataSource) {
       dataSourceProp = this.options.snakeCase ? { data_source: dataSource } : { dataSource };
     }
-      
+
     const sqlOption = this.options.snakeCase
       ? {
         sql_table: table,
@@ -126,7 +126,7 @@ export abstract class BaseSchemaFormatter {
       cube: tableSchema.cube,
       ...sqlOption,
       ...dataSourceProp,
-      
+
       joins: tableSchema.joins
         .map((j) => ({
           [j.cubeToJoin]: {
@@ -164,7 +164,7 @@ export abstract class BaseSchemaFormatter {
             type: 'count',
           },
         }),
-        
+
       ...(this.options.snakeCase
         ? Object.fromEntries(
           Object.entries(contextProps).map(([key, value]) => [toSnakeCase(key), value])