+++ rust/cubestore/cubestore/src/queryplanner/optimizations/mod.rs
@@ -1,3 +1,8 @@
+mod check_memory;
+mod distributed_partial_aggregate;
+mod prefer_inplace_aggregates;
+pub mod rewrite_plan;
+
 use crate::cluster::Cluster;
 use crate::queryplanner::optimizations::distributed_partial_aggregate::{
     add_limit_to_workers, push_aggregate_to_workers,
@@ -5,6 +10,8 @@ use crate::queryplanner::optimizations::distributed_partial_aggregate::{
 use crate::queryplanner::optimizations::prefer_inplace_aggregates::try_switch_to_inplace_aggregates;
 use crate::queryplanner::planning::CubeExtensionPlanner;
 use crate::queryplanner::serialized_plan::SerializedPlan;
+use crate::util::memory::MemoryHandler;
+use check_memory::add_check_memory_exec;
 use datafusion::error::DataFusionError;
 use datafusion::execution::context::{ExecutionContextState, QueryPlanner};
 use datafusion::logical_plan::LogicalPlan;
@@ -13,30 +20,33 @@ use datafusion::physical_plan::{ExecutionPlan, PhysicalPlanner};
 use rewrite_plan::rewrite_physical_plan;
 use std::sync::Arc;
 
-mod distributed_partial_aggregate;
-mod prefer_inplace_aggregates;
-pub mod rewrite_plan;
-
 pub struct CubeQueryPlanner {
     cluster: Option<Arc<dyn Cluster>>,
     serialized_plan: Arc<SerializedPlan>,
+    memory_handler: Arc<dyn MemoryHandler>,
 }
 
 impl CubeQueryPlanner {
     pub fn new_on_router(
         cluster: Arc<dyn Cluster>,
         serialized_plan: Arc<SerializedPlan>,
+        memory_handler: Arc<dyn MemoryHandler>,
     ) -> CubeQueryPlanner {
         CubeQueryPlanner {
             cluster: Some(cluster),
             serialized_plan,
+            memory_handler,
         }
     }
 
-    pub fn new_on_worker(serialized_plan: Arc<SerializedPlan>) -> CubeQueryPlanner {
+    pub fn new_on_worker(
+        serialized_plan: Arc<SerializedPlan>,
+        memory_handler: Arc<dyn MemoryHandler>,
+    ) -> CubeQueryPlanner {
         CubeQueryPlanner {
             serialized_plan,
             cluster: None,
+            memory_handler,
         }
     }
 }
@@ -54,14 +64,18 @@ impl QueryPlanner for CubeQueryPlanner {
             })])
             .create_physical_plan(logical_plan, ctx_state)?;
         // TODO: assert there is only a single ClusterSendExec in the plan.
-        finalize_physical_plan(p)
+        finalize_physical_plan(p, self.memory_handler.clone())
     }
 }
 
 fn finalize_physical_plan(
     p: Arc<dyn ExecutionPlan>,
+    memory_handler: Arc<dyn MemoryHandler>,
 ) -> Result<Arc<dyn ExecutionPlan>, DataFusionError> {
     let p = rewrite_physical_plan(p.as_ref(), &mut |p| try_switch_to_inplace_aggregates(p))?;
     let p = rewrite_physical_plan(p.as_ref(), &mut |p| push_aggregate_to_workers(p))?;
+    let p = rewrite_physical_plan(p.as_ref(), &mut |p| {
+        add_check_memory_exec(p, memory_handler.clone())
+    })?;
     rewrite_physical_plan(p.as_ref(), &mut |p| add_limit_to_workers(p))
 }